Clean up maker code, address PR suggestions
